Create README.md
Brief overview of the project and how it works, along with the historic trading reccomendations, and the program's win rate. I have also included an example program output to show what the code displays on a particular day it was run, making a prediction about the S&P500 price the next day.